complete and test check owner ID for edit/del restaurant info

+++ controllers/restaurant.js
@@ -60,25 +60,35 @@ async function createRestaurant(req, res) {
 }
 
 async function editRestaurant(req, res) {
-  try {
-    const data = await modelRestaurant.editRestaurant(
-      req.params.restId,
-      req.body
-    );
-    res.json(data);
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ errorMsg: err.message });
+  const restdata = await modelRestaurant.getRestaurantById(req.params.restId);
+  if (!restdata.owner || restdata.owner != req.user.id) {
+    return res.status(401).json("Unauthorized");
+  } else {
+    try {
+      const data = await modelRestaurant.editRestaurant(
+        req.params.restId,
+        req.body
+      );
+      res.json(data);
+    } catch (err) {
+      console.error(err);
+      res.status(500).json({ errorMsg: err.message });
+    }
   }
 }
 
 async function deleteRestaurant(req, res) {
-  try {
-    await modelRestaurant.deleteRestaurant(req.params.restId);
-    res.json("data has been deleted.");
-    // res.redirect('/');
-  } catch (err) {
-    console.error(err);
-    res.status(500).json({ errorMsg: err.message });
+  const restdata = await modelRestaurant.getRestaurantById(req.params.restId);
+  if (!restdata.owner || restdata.owner != req.user.id) {
+    return res.status(401).json("Unauthorized");
+  } else {
+    try {
+      await modelRestaurant.deleteRestaurant(req.params.restId);
+      res.json("data has been deleted.");
+      // res.redirect('/');
+    } catch (err) {
+      console.error(err);
+      res.status(500).json({ errorMsg: err.message });
+    }
   }
 }
